frontend/telegram_bot/requirements.txt
python-telegram-bot
httpx
pyyaml